Rename file footer to footer-section
(cherry picked from commit 68f1545af426f5b1b80a3c80912ac747a6602a9f)
