+++ src/DetailsView/reports/components/report-sections/footer-section.tsx
